+++ CHANGELOG.md
@@ -6,7 +6,9 @@ This project now adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 - Removed built in phpunit integration
-- Removed the `named-only` command line option
+- Removed the `--named-only` command line option
+- Added the `--file-extension` command line option
+- Added the `--ignore` command line option
 
 ## [1.0.0] - 2020-07-06